requirements.txt
alembic==1.10.2
astroid==2.15.0
attrs==22.2.0
autopep8==2.0.2
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.2
dill==0.3.6
flasgger==0.9.5
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mistune==2.0.5
mysqlclient==2.1.1
paramiko==2.12.0
pep8==1.7.1
pkgutil_resolve_name==1.3.10
platformdirs==3.1.1
pycodestyle==2.10.0
pycparser==2.21
pylint==2.17.1
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.7
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
